Move ImmutableObjectUtils to mapper module
